+++ client/src/components/LoggedIn/OrderMain.js
@@ -24,7 +24,7 @@ const OrderMain = () => {
         <>
           <Header>Select from a quote below or create a new quote</Header>
           <Orders />
-          Hello Landon that feels goooooood Hello again
+          
         </>
       )
     }
