+++ client/src/components/LoggedIn/Orders.js
@@ -5,47 +5,11 @@ import { Header } from "semantic-ui-react"
 import Order from "./Order"
 import OrderForm from "./OrderForm"
 import { AuthContext } from "../../providers/AuthProvider"
+import { OrdersContext } from "../../providers/OrdersProvider"
 
 const Orders = (props) => {
 
-  const [orders, setOrders] = useState([])
-  const [loading, setLoading] = useState(true)
-  const [error, setError] = useState(false)
-
-  const {setLiveOrder} = useContext(AuthContext)
-
-  useEffect(()=>{
-    getOrders()
-  },[])
-
-  const getOrders = async () => {
-    try { 
-      let res = await Axios.get("/api/orders")
-      setOrders(res.data)
-
-    } catch (err) {
-      console.log(err)
-      setError(true)
-
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  const createOrder = async (newOrder) => {
-    console.log(newOrder)
-    try{
-      // debugger
-      let res = await Axios.post(`/api/orders`,newOrder)
-      console.log("res: ",res.data)
-      setOrders([...orders, res.data])
-      console.log(res.data.id)
-      setLiveOrder(res.data.id)
-    }
-    catch(err){
-      console.log(err)
-    }
-  }
+  const {createOrder,loading,error,orders, hello} = useContext(OrdersContext)
 
   const renderOrders = ()=>{
